Support miniature cat-shaped printers
This swaps out the ECS/POS library for a custom BLE implementation,
reflecting a real change in underlying hardware. It is experimental: the
smaller printer is shared, and has significant drawbacks when compared
to the original larger device (slower print speed, manual slicing, etc).
However, its extreme portability is interesting as a proof-of-concept.

The protocol implementation in the `src/cat` folder is based off this
project: https://github.com/rbaron/catprinter. There exist very few
libraries for working with the protocol (it is undocumented, after all)
and even fewer that work with text instead of images (because the device
appears to only natively support images and manually rendering text is a
pain). I plan to refactor my code out to be usable for others as well in
the near future—keep an eye on PyPi.

Cat-shaped hardware provided by: https://www.frankchiarulli.com/
Environment to build fun things provided by: https://www.recurse.com/